# api/openapi.yaml
openapi: 3.0.3
info:
  title: Geek Budget - OpenAPI 3.0
  contact:
    email: ilya.korolev@outlook.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 0.0.1
...
paths:
...
  /v1/accounts/{accountId}/history:
    get:
      tags:
        - accounts
      summary: return list of dates when this account was used in some transaction
      operationId: getAccountHistory
      parameters:
        - name: "accountId"
          in: "path"
          description: "ID of account"
          required: true
          schema:
            type: "string"
            format: "uuid"
            example: "123e4567-e89b-12d3-a456-426614174000"
      responses:
        "200":
          description: list of dates when this account was used in some transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountHistory"
  /v1/accounts:
    get:
      tags:
        - accounts
      summary: get all accounts
      operationId: getAccounts
      responses:
        "200":
          description: accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Account"
    post:
      tags:
        - accounts
      summary: create new account
      operationId: createAccount
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccountNoID"
      responses:
        "200":
          description: created account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
  /v1/accounts/{id}:
    get:
      tags:
        - accounts
      summary: get account
      operationId: getAccount
      parameters:
        - name: "id"
          in: "path"
          description: "ID of the account"
          required: true
          schema:
            type: "string"
            format: "uuid"
            example: "123e4567-e89b-12d3-a456-426614174000"
      responses:
        "200":
          description: account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
    put:
      tags:
        - accounts
      summary: update account
      operationId: updateAccount
      parameters:
        - name: "id"
          in: "path"
          description: "ID of the account"
          required: true
          schema:
            type: "string"
            format: "uuid"
            example: "123e4567-e89b-12d3-a456-426614174000"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccountNoID"
      responses:
        "200":
          description: updated account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
    delete:
      tags:
        - accounts
      summary: delete account
      operationId: deleteAccount
      parameters:
        - name: "id"
          in: "path"
          description: "ID of the account"
          required: true
          schema:
            type: "string"
            format: "uuid"
            example: "123e4567-e89b-12d3-a456-426614174000"
      responses:
        "200":
          description: no body
  /v1/accounts/{id}/image:
    post:
      tags:
        - accounts
      summary: Upload account image
      operationId: uploadAccountImage
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "string"
            format: "uuid"
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "200":
          description: status of the upload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
    delete:
      tags:
        - accounts
      summary: delete account image
      operationId: deleteAccountImage
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "string"
            format: "uuid"
      responses:
        "200":
          description: updated account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
...
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Entity:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
      required:
        - id
...
    BankAccountInfo:
      type: object
      properties:
        accountId:
          type: string
          example: "12345"
        bankId:
          type: string
          example: "12345"
        balances:
          description: >-
            List of balances for this account. It's an array since one account could hold multiple
            currencies, for example, cash account could hold EUR, USD and CZK. Or one bank account could
            hold multiple currencies.
          type: array
          items:
            type: object
            properties:
              currencyId:
                type: string
                format: uuid
              openingBalance:
                type: number
                format: double
                example: 123.0
              closingBalance:
                type: number
                format: double
                example: 321.0
    AccountNoID:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        type:
          enum:
            - expense
            - income
            - asset
        bankInfo:
          $ref: "#/components/schemas/BankAccountInfo"
        showInDashboardSummary:
          type: boolean
          description: "If true, show this account in dashboard summary."
          default: true
        image:
          type: string
          description: "ID of the account image"
      required:
        - name
        - type
    Account:
      type: object
      allOf:
        - $ref: "#/components/schemas/Entity"
        - $ref: "#/components/schemas/AccountNoID"
      required:
        - showInDashboardSummary
...
    AccountHistory:
      type: array
      items:
        type: string
        format: date-time
        example: "2020-01-01T00:00:00Z"
...