# backend/pkg/generated/goserver/model_account.go
// Code generated by OpenAPI Generator (https://openapi-generator.tech); DO NOT EDIT.

/*
 * Geek Budget - OpenAPI 3.0
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * API version: 0.0.1
 * Contact: ilya.korolev@outlook.com
 */

package goserver

type Account struct {
	Id string `json:"id"`

	Name string `json:"name"`

	Description string `json:"description,omitempty"`

	Type string `json:"type"`

	BankInfo BankAccountInfo `json:"bankInfo,omitempty"`

	// If true, show this account in dashboard summary.
	ShowInDashboardSummary bool `json:"showInDashboardSummary"`

	// ID of the account image
	Image string `json:"image,omitempty"`
}

type AccountInterface interface {
	GetId() string
	GetName() string
	GetDescription() string
	GetType() string
	GetBankInfo() BankAccountInfo
	GetShowInDashboardSummary() bool
	GetImage() string
}

func (c *Account) GetId() string {
	return c.Id
}
func (c *Account) GetName() string {
	return c.Name
}
func (c *Account) GetDescription() string {
	return c.Description
}
func (c *Account) GetType() string {
	return c.Type
}
func (c *Account) GetBankInfo() BankAccountInfo {
	return c.BankInfo
}
func (c *Account) GetShowInDashboardSummary() bool {
	return c.ShowInDashboardSummary
}
func (c *Account) GetImage() string {
	return c.Image
}

// AssertAccountRequired checks if the required fields are not zero-ed
func AssertAccountRequired(obj Account) error {
	elements := map[string]interface{}{
		"id":                     obj.Id,
		"name":                   obj.Name,
		"type":                   obj.Type,
		"showInDashboardSummary": obj.ShowInDashboardSummary,
	}
	for name, el := range elements {
		if isZero := IsZeroValue(el); isZero {
			return &RequiredError{Field: name}
		}
	}

	if err := AssertBankAccountInfoRequired(obj.BankInfo); err != nil {
		return err
	}
	return nil
}

// AssertAccountConstraints checks if the values respects the defined constraints
func AssertAccountConstraints(obj Account) error {
	if err := AssertBankAccountInfoConstraints(obj.BankInfo); err != nil {
		return err
	}
	return nil
}